Blink-rate Detection/requirements.txt
# Eye Blink Rate Detector – Dependencies
# Install with: pip install -r requirements.txt

# Core computer vision
opencv-python>=4.8.0
mediapipe>=0.10.0

# Document rendering
PyMuPDF>=1.23.0          # fitz – PDF rendering

# Image handling (for Tkinter display)
Pillow>=10.0.0

# Charts
matplotlib>=3.8.0

# DOCX support
python-docx>=1.1.0       # For DOCX metadata (LibreOffice handles conversion)

# Optional: VLC-based video playback (alternative to OpenCV player)
# python-vlc>=3.0.0      # Uncomment if VLC is installed on your system

# Audio support for video playback
pygame>=2.0.0
